refactor(api): package dependency providers
